+++ packages/nhsuk-frontend/src/nhsuk/components/panel/macro-options.mjs
@@ -32,13 +32,13 @@ export const params = {
   },
   text: {
     type: 'string',
-    required: false,
+    required: true,
     description:
       'If `html` is set, this is not required. Text to use within the panel content. If `html` is provided, the `text` option will be ignored.'
   },
   html: {
     type: 'string',
-    required: false,
+    required: true,
     description:
       'If `text` is set, this is not required. Text to use within the panel content. If `text` is provided, the `html` option will be ignored.'
   },